backend/requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg[binary]>=3.2,<4.0
python-dotenv>=1.0,<2.0
django-cors-headers>=4.6,<5.0
beautifulsoup4>=4.12,<5.0
requests>=2.32,<3.0
PyPDF2>=3.0,<4.0
supabase>=2.10,<3.0
numpy>=1.26,<2.0
pandas>=2.2,<3.0
scikit-learn>=1.4,<2.0



